# Documentacion/CraftQuest_OpenAPI_MVP_v4.yaml
openapi: 3.0.3
info:
  title: CraftQuest API
  version: "4.0.0"
  description: >
    API MVP completo v4. Las respuestas se identifican por AnswerOptionId.
    A/B/C/D son etiquetas visuales generadas por intento y guardadas como snapshot.
servers:
  - url: https://api.craftquestai.com/api
    description: Producción
  - url: https://localhost:7080/api
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /study-materials:
    get:
      tags: [StudyMaterials]
      summary: Listar materiales del usuario
      parameters:
        - name: skip
          in: query
          schema: { type: integer, default: 0 }
        - name: take
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Lista de materiales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterialSummaryDto"
    post:
      tags: [StudyMaterials]
      summary: Subir PDF o DOCX
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                title: { type: string }
      responses:
        "202":
          description: Material aceptado para procesamiento
...
  /study-materials/{studyMaterialId}:
    get:
      tags: [StudyMaterials]
      summary: Detalle del material (paginas, secciones, estado)
      parameters:
        - $ref: "#/components/parameters/StudyMaterialId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterialDetailDto"
  /study-materials/{studyMaterialId}/extracted-text:
    patch:
      tags: [StudyMaterials]
      summary: Guardar texto revisado manualmente
      parameters:
        - $ref: "#/components/parameters/StudyMaterialId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudyMaterialExtractedTextRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterialDetailDto"
...
  /study-materials/{studyMaterialId}/selection:
    patch:
      tags: [StudyMaterials]
      summary: Actualizar rango de paginas y tema
      parameters:
        - $ref: "#/components/parameters/StudyMaterialId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudyMaterialSelectionRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterialDetailDto"
  /study-materials/{studyMaterialId}/generate/estimate:
    post:
      tags: [StudyMaterials]
      summary: Estimar creditos y preguntas importables
      parameters:
        - $ref: "#/components/parameters/StudyMaterialId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizGenerationParameters"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizGenerationEstimateDto"
  /study-materials/{studyMaterialId}/generate:
    post:
      tags: [StudyMaterials]
      summary: Iniciar generacion de cuestionario con IA
      parameters:
        - $ref: "#/components/parameters/StudyMaterialId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizGenerationParameters"
      responses:
        "202":
          description: Job de generacion creado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StudyMaterialId:
      name: studyMaterialId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    StudyMaterialSummaryDto:
      type: object
      properties:
        studyMaterialId: { type: string, format: uuid }
        title: { type: string }
        fileType: { type: string }
        processingStatus: { type: string }
        needsOcr: { type: boolean }
        pageCount: { type: integer, nullable: true }
        wordCount: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    StudyMaterialDetailDto:
      type: object
      properties:
        studyMaterialId: { type: string, format: uuid }
        title: { type: string }
        fileType: { type: string }
        processingStatus: { type: string }
        needsOcr: { type: boolean }
        requiresTextReview: { type: boolean }
        editedExtractedText: { type: string, nullable: true }
        pageCount: { type: integer, nullable: true }
        wordCount: { type: integer, nullable: true }
        estimatedMaxQuestions: { type: integer }
        pages:
          type: array
          items:
            $ref: "#/components/schemas/StudyMaterialPageDto"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/StudyMaterialSectionDto"
    StudyMaterialPageDto:
      type: object
      properties:
        pageNumber: { type: integer }
        previewText: { type: string, nullable: true }
        wordCount: { type: integer }
        extractionQuality: { type: string }
    StudyMaterialSectionDto:
      type: object
      properties:
        title: { type: string }
        pageFrom: { type: integer }
        pageTo: { type: integer }
    UpdateStudyMaterialExtractedTextRequest:
      type: object
      required: [extractedText]
      properties:
        extractedText: { type: string }
    UpdateStudyMaterialSelectionRequest:
      type: object
      properties:
        pageFrom: { type: integer }
        pageTo: { type: integer }
        topic: { type: string, nullable: true }
    QuizGenerationParameters:
      type: object
      properties:
        targetQuizId: { type: string, format: uuid, nullable: true }
        questionCount: { type: integer }
        language: { type: string }
        difficulty: { type: string }
        allowedQuestionTypes:
          type: array
          items: { type: string }
        topicFocus: { type: string, nullable: true }
        pageFrom: { type: integer }
        pageTo: { type: integer }
    QuizGenerationEstimateDto:
      type: object
      properties:
        creditsRequired: { type: integer }
        aiCreditsAvailable: { type: integer }
        estimatedImportableQuestions: { type: integer }
        wordsInScope: { type: integer }
...